+++ app/routers/tilerace/__init__.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter
 
 from .completions import router as completions_router
 from .controls import router as controls_router
+from .discord_setup import router as discord_setup_router
 from .events import router as events_router
 from .generate import router as generate_router
 from .osrs_ref import router as osrs_ref_router
@@ -25,3 +26,4 @@ router.include_router(controls_router)
 router.include_router(completions_router)
 router.include_router(rolls_router)
 router.include_router(sabotage_router)
+router.include_router(discord_setup_router)